mmaintegration: move deprecated logical model definitions to test scope
The `storeCPURateCapacityInput` type, `computeStoreByteSizeCapacity`,
`cpuCapacityFloorPerStore`, and `computeCPUCapacityWithCap` are only
referenced by tests. This change moves them into
`deprecated_logical_model_test.go` and deletes the non-test file.

The `maxCPUAmplification` constant is the exception — it is used by
`computePhysicalCPU` in `physical_model.go`. This change relocates it
there.
